resume-analyzer-backend/requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
docx2txt
sentence-transformers
requests
python-dotenv